refactor: modularize email utils
