+++ utils/email/email.py
@@ -2,36 +2,11 @@ import os
 
 import resend
 from dotenv import load_dotenv
-from ninja import Schema
+
+from utils.email.other import EMAIL_TEMPLATE
+from utils.email.schemas import EmailIn, EmailResponse
 
 load_dotenv()
-
-
-class EmailIn(Schema):
-    """
-    Email in schema
-    """
-
-    name: str
-    email: str
-    motivation: str
-    message: str
-
-
-class EmailResponse(Schema):
-    """
-    Email response schema
-    """
-
-    email_id: str
-
-
-email_template = """
-    <div>
-        <h2>You have received a message from {email}</h2>
-        <p>The message reads as follows: <i>"{message}"</i></p>
-    </div>
-"""
 
 
 class EmailHandler:
@@ -62,7 +37,7 @@ class EmailHandler:
             "from": "SPM <onboarding@resend.dev>",
             "to": ["sergio.pm.developer@gmail.com"],
             "subject": f"SPM: {email.motivation}",
-            "html": email_template.format(email=email.email, message=email.message),
+            "html": EMAIL_TEMPLATE.format(email=email.email, message=email.message),
         }
 
         email_sent = resend.Emails.send(params)
