+++ utils/email/other.py
@@ -1,0 +1,6 @@
+EMAIL_TEMPLATE = """
+    <div>
+        <h2>You have received a message from {email}</h2>
+        <p>The message reads as follows: <i>"{message}"</i></p>
+    </div>
+"""
